% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options-package.R
\name{ip_options}
\alias{ip_options}
\alias{ip_get_options}
\alias{ip_get_option}
\alias{get_pkg_options}
\title{Package options}
\usage{
ip_options(...)

ip_get_options(pattern = NULL)

ip_get_option(x)
}
\arguments{
\item{...}{set package options, syntax identical to \code{\link[=options]{options()}}}

\item{pattern}{to retrieve multiple options (as a list) with a shared pattern}

\item{x}{name of the specific option to retrieve}
}
\value{
\code{ip_options()} and \code{ip_get_options()} return a named list of option
values; \code{ip_get_option()} returns the value of the single requested option.
}
\description{
These options are best set via \code{\link[=ip_options]{ip_options()}} and queried via \code{\link[=ip_get_option]{ip_get_option()}}.
However, the base functions \code{\link[=options]{options()}} and \code{\link[=getOption]{getOption()}} work as well but require
an \code{isoprocessor2.} prefix (the package name and a dot) for the option name. Setting
an option to a value of \code{NULL} means that the default is used. \code{\link[=ip_get_options]{ip_get_options()}}
is available as an additional convenience function to retrieve a subset of options
with a regular expression pattern.
}
\section{Functions}{
\itemize{
\item \code{ip_options()}: set/get option values

\item \code{ip_get_options()}: get a subset of option values that fit a pattern

\item \code{ip_get_option()}: retrieve the current value of one option (option must be defined for the package)

}}
\section{Options for the isoprocessor2 package}{

\itemize{
\item \code{debug}: turn on debug mode
\item \code{auto_use_ansi}: whether to automatically enable correct rendering of stylized (ansi) output in HTML reports from notebooks that call \code{library(isoprocessor2)}. Can be turned off by calling \code{isoprocessor2::ip_options(auto_use_ansi = FALSE)} \strong{before} calling \code{library(isoprocessor2)}.
}
}

\examples{
# All default options
ip_get_options()

}
