% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoprocessor2-package.R
\docType{package}
\name{isoprocessor2-package}
\alias{isoprocessor2}
\alias{isoprocessor2-package}
\title{isoprocessor2: Process Stable Isotope Data}
\description{
Data processing and reduction pipelines for the stable isotope data read by the isoreader2 package, including peak detection, calibration, standardization, and visualization of isotope ratio mass spectrometry (IRMS) data.
}
\details{
Resources:
\itemize{
\item Website for the isoprocessor2 package: \url{https://isoprocessor2.isoverse.org}
\item Package options: \link{ip_options}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isoprocessor2.isoverse.org/}
  \item \url{https://github.com/isoverse/isoprocessor2}
  \item Report bugs at \url{https://github.com/isoverse/isoprocessor2/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})

Authors:
\itemize{
  \item Sebastian Kopf \email{sebastian.kopf@colorado.edu} (\href{https://orcid.org/0000-0002-2044-0201}{ORCID})
}

}
