% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot detected peaks on a chromatogram}
\usage{
plot_peaks(time, intensity, peaks, show_background = TRUE)
}
\arguments{
\item{time}{Numeric vector. Time in seconds.}

\item{intensity}{Numeric vector. Intensity in volts.}

\item{peaks}{Tibble from \code{integrate_peaks()} or \code{find_peaks()}.}

\item{show_background}{Logical. If TRUE, draw background lines. Default TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Plot detected peaks on a chromatogram
}
